// src/app/models/part.ts
export class Part {
    AutomotivePartID: number;
    PartNameEnglish: string;
    PartNameArabic: string;
    ItemCode: string;
    DescriptionEng: string;
    DescriptionArabic: string;
    MainCategoryNameEng: string;
    MainCategoryNameArabic: string;
    SubCategoryNameEng: string;
    SubCategoryNameArabic: string;
    SubSubCategoryNameEng: string;
    SubSubCategoryNameArabic: string;
    MainCategoryID: number;
    SubCategoryID: number;
    SubSubCategoryID: number;
    UnitQuantity : number;
    PurchasePrice: number;
    SellingPrice: number;
    HighestSellingPrice: number;
    PartPrice: number;
    OEMNumber: string;
    VendorName: string;
    Tax: number;
    NetPrice: number ;
    SiteID: number;
    ManufacturerID: number;
    OrigionID: number;
    CreatedBy: number;
    ModifiedBy: number;
    LaborCost: number;
    LaborTime: number;
    WorkshopID: number;
    IsModified: boolean;
    SubstitutePart: string;
    LogoImageURL: string;
    IsNew: boolean = false;
    IsUsed: boolean;
    VendorPrice: number;
    Address1: string;
    CPPhone: number;
    ManufacturerNameEng: string;
    Quantity: number;
    IsChecked: boolean;
    TotalPages: number;
    Hours: number;
    Minutes: number;
    BrandName: any;
    BrandID: number;
    OriginName: any;
    DiscountPercentage: number;
    JobPartStatusID: number;
    JobPartAlternateID: number
    StockQuantity: number;
    Viscosity: string;
    UnitTypeID:number;
    UnitTypeName: string;
    MainPartID: number;
    ItemNameArabic: string; 
    ItemNameEng: string; 
    JobPartID: number;
    IsInclude :boolean;
}